Moves dynamic table resizing into ContentPanelStore
